% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_other_rels}
\alias{example_data_other_rels}
\title{OSM infrastructure data used in SOTM2022 workshop & figures for SOTM2022.}
\format{
A data frame of OSM other relations within a 2.5km radius of Leeds
City Centre.
}
\source{
{\href{https://github.com/udsleeds/openinfra/blob/main/test-code/generate_example_data.R}{generate_example_data.R}}
}
\usage{
example_data_other_rels
}
\description{
OSM infrastructure data used in SOTM2022 workshop & figures for SOTM2022.
}
\examples{
library(sf)
colnames(example_data_other_rels)
}
\keyword{dataset}
