% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_pois}
\alias{example_data_pois}
\title{OSM points data used in the creation of transport infrastructure data packs.}
\format{
A dat frame of OSM ways.
}
\source{
{\href{https://github.com/udsleeds/openinfra/blob/main/test-code/generate_example_data.R}{generate_example_data.R}}
}
\usage{
example_data_pois
}
\description{
OSM points data used in the creation of transport infrastructure data packs.
}
\examples{
library(sf)
example_data_pois
}
\keyword{dataset}
