% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extended_data}
\alias{extended_data}
\title{OSM infrastructure data used in creation of transport infrastructure data packs.}
\format{
A data frame with 42,779 features described by 28 variables.
}
\source{
{\href{https://github.com/udsleeds/openinfra/blob/main/test-code/generate_example_data.R}{generate_example_data.R}}
}
\usage{
extended_data
}
\description{
Contains all columns required for IM function, recode_max_speed, recode_road_class.
}
\examples{
library(sf)
extended_data   # Lazy Loading. Data becomes visible as soon as it is requested. 
}
\keyword{dataset}
