% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_active_walk.R
\name{oi_active_walk}
\alias{oi_active_walk}
\title{Function to recategorise OSM infrastrcutre based on pedestrian walkability.}
\usage{
oi_active_walk(osm_sf, remove=FALSE)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} function.
}}

\item{remove}{\itemize{
\item Boolean - If TRUE, features that do not form part of a
walkable network are removed from the \code{osm_sf} before being returned
}}
}
\value{
an sf object with oi_walk column added.
}
\description{
Determins whether or not a piece of OSM infrastructure is walkable or not
depending on whether the feature is included in the default walking network
defined by \href{https://github.com/ropensci/osmextract}{\code{osmextract}}
}
\details{
This function mimics the walking filtering steps applied by the
\href{https://rdrr.io/cran/osmextract/man/oe_get_network.html}{\code{oe_get_network}}
function, part of the
\href{https://rdrr.io/cran/osmextract/}{\code{osmextract}} package.

Note: the \code{osm_sf} must contain the following tags: \code{c("highway", "foot", "access", "service")}
}
\examples{
data = example_data
example_output = oi_active_walk(data)
}
