% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_bicycle_parking.R
\name{oi_bicycle_parking}
\alias{oi_bicycle_parking}
\title{Identify bicycle parking places from OSM}
\usage{
oi_bicycle_parking(osm_sf, remove = FALSE)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} function.
}}

\item{remove}{\itemize{
\item if \code{TRUE}, removes all nodes that are not suitable for
bicycle parking.
}}
}
\value{
\itemize{
\item The provided \code{osm_sf} is returned with an additional column,
\code{oi_cycle_parking} containing the value \code{"yes"} if this feature allows
bicycle parking.
}
}
\description{
This function re-categorises OSM data, adding a column to an OSM network
\code{oi_cycle_parking} with value \code{"yes"} if a node/way within the OSM network has
facilities for bicycle parking.
}
\details{
\itemize{
\item Note: the input \code{osm_sf} must containt the following tags:
\code{c("amenity")}.
\item Note that as most (if not all) amenity tags for bicycle
parking are assigned to nodes, the \code{osm_sf} must also contain nodes. (the
layer \code{"points"} must be requested when using \code{osmextract::oe_get()})
}
}
\examples{
data = example_data_pois
example_output = oi_bicycle_parking(data)
example_output = example_output \%>\% dplyr::filter(oi_cycle_parking == "yes")
plot(example_output["oi_cycle_parking"])
}
