% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_cycle_crossings.R
\name{oi_cycle_crossings}
\alias{oi_cycle_crossings}
\title{Function to get OSM data based on cyclist crossing infrastructure.}
\usage{
oi_bicycle_parking(osm_sf, remove=FALSE)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} function.
}}

\item{remove}{\itemize{
\item Boolean - If TRUE, features that do not form part of cycle
crossings are removed from the \code{osm_sf} before being returned.
}}
}
\value{
an \code{sf} object with openinfra_cycle_crossings column added.
}
\description{
Function to get OSM data based on cyclist crossing infrastructure.
}
\details{
\itemize{
\item Note: the \code{osm_sf} must contain the following tags:
\code{c("cycleway", "crossing", "bicycle", "crossing_island", "crossing:island", "crossing:ref", "crossing_ref")}
\itemize{
\item Note: The \code{osm_sf} must contain linestring geometries rather than nodes.
Further motivation behind this function can be found
\href{https://github.com/udsleeds/openinfra/issues/126}{here}.
}
}
}
\examples{
data = example_data
example_output = oi_cycle_crossings(data)
plot(example_output["openinfra_cycle_crossings"])
  
}
