% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_cycle_separation.R
\name{oi_cycle_separation}
\alias{oi_cycle_separation}
\title{Function to recategorise OSM data based on the type of cycle infrastructure.}
\usage{
oi_cycle_infra(osm_sf, remove=FALSE)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} function.
}}

\item{remove}{\itemize{
\item Boolean, FALSE by default, if TRUE, removes ways that are not
part of any dedicated cycling infrastructure.
}}
}
\value{
The provided \code{osm_sf} is returned with an additional column,
\code{openinfra_cycle_infa}, which indicates if a way is part of cycling
infrastructure, and the level of protection provided by the infrastructure
("mixed traffic", "cycle lanes", "protected cycling space")
}
\description{
Function to recategorise OSM data based on the type of cycle infrastructure.
}
\details{
This function recetegorises OSM data on cycling infrastrcutre to one of the
protection categories defined within the LTN1/20 guidance (Mixed Traffic |
Cycle Lanes | Protected Cycling Space).

\strong{Note}: The \code{osm_sf} must contain the following tags: c("cycleway",
"cycleway_left", "cycleway_right", "cycleway_both")
}
\examples{
library(sf)
data = example_data
example_output = oi_cycle_separation(data)
plot(example_output["openinfra_cycle_infra"]) 
}
