% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_im_flush_kerb.R
\name{oi_im_flush_kerb}
\alias{oi_im_flush_kerb}
\title{Assesses whether a flush kerb is compliant with Inclusive Mobility guidance.}
\usage{
oi_im_flush_kerb(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} package.
}}
}
\value{
The \code{osm_sf} data frame is returned with additional
columns \code{openinfra_im_tactile_paving}, assessing the presence of tactile
paving with values \code{c("yes", "no")}, and
\code{openinfra_im_flush_kerb}, which assesses whether or not there is a flush
kerb compliant with the latest
\href{https://tinyurl.com/IM-guide-link}{Inclusive Mobility} Guidance.
}
\description{
This function is used to assess if a flush kerb is complicit with current
\href{https://tinyurl.com/IM-guide-link}{Inclusive Mobility} guidance. The
function first assesses the presence of tactile paving, which should be
present for a compliant flush kerb, before assessing the kerb type.
}
\details{
Note: \code{osm_sf} must contain the tags: \code{c("kerb", "tactile_paving")}.
}
\examples{
data = example_data
output = oi_im_flush_kerb(data)
output = output \%>\% dplyr::filter(! is.na(openinfra_im_flush_kerb))
plot(output[,"openinfra_im_flush_kerb"])
}
