% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_im_pavement_width.R
\name{oi_im_pavement_width}
\alias{oi_im_pavement_width}
\title{Function to assess pedestrian infrastructure widths.}
\usage{
oi_im_pavement_width(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} package.
}}
}
\value{
the \code{osm_sf} is returned with additional columns \code{openinfra_im_width}
and \code{openinfra_im_width_est} with recategorised way width and estimated
width value ranges appended respectively.
}
\description{
Function to assess pedestrian infrastructure widths.
}
\details{
Note: the \code{osm_sf} must contain the following tags: \code{c("width",  "est_width")}
}
\examples{
data = example_data
output = oi_im_pavement_width(data)
plot(output["openinfra_im_width_est"])
# Uncomment below to plot other columns
# plot(output["openinfra_im_width"])
}
