% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_im_pedestrian_infra.R
\name{oi_im_pedestrian_infra}
\alias{oi_im_pedestrian_infra}
\title{Function to assess pedestrian infrastructure data.}
\usage{
oi_im_pedestrian_infra(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} package.
}}
}
\value{
The \code{osm_sf} data frame is returned with additional columns added:
\code{openinfra_im_footway} which assesses the presence of footways with values
\code{c("footway", "no")}. \code{openinfra_im_footpath} which assesses the presence
of footpaths with values \code{c("footpath", "no")}. \code{openinfra_im_crossing}
which assesses the presence and type of pedestrian crossing with values
\code{c("give-way crossing", "signal-controlled crossing", "unknown crossing type", "no")}. \code{openinfra_im_footway_imp} which assesses
the presence of implied footways not implicitly tagged with values
\code{c("implied footway", "no")}.
}
\description{
Function to assess pedestrian infrastructure data.
}
\details{
Note: the \code{osm_sf} must contain the following tags: \code{c( "footway", "sidewalk", "cycleway", "foot", "segregated", "access",  "crossing", "footway")}
}
\examples{
data = example_data
output = oi_im_pedestrian_infra(data)
plot(output["openinfra_im_footway"])
# Uncomment below to plot other columns
#plot(output["openinfra_im_footpath"])
#plot(output["openinfra_im_crossing"])
#plot(output["openinfra_im_footway_imp"])
}
