% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_im_surfaces.R
\name{oi_im_surfaces}
\alias{oi_im_surfaces}
\title{Function to assess surfaces of OSM ways.}
\usage{
oi_im_surfaces(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained using the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} package.
}}
}
\value{
the \code{osm_sf} is returned with additional columns
\code{openinfra_im_surface_paved} and \code{openinfra_im_surface} containing values
on presence of surface paving, and whether or not the surface is even
respectively.
}
\description{
Function to assess surfaces of OSM ways.
}
\details{
Note: the \code{osm_sf} must contain the following tags: \code{c("surface",  "smoothness")}
}
