% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_im_tactile_paving.R
\name{oi_im_tactile_paving}
\alias{oi_im_tactile_paving}
\title{Assess the presence of tactile paving from OSM data.}
\usage{
oi_im_tactile_paving(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} package.
}}
}
\value{
The \code{osm_sf} data frame is returned with additional
\code{openinfra_im_tactile_paving} column, containing values \code{c("yes", "no")}
}
\description{
Assess the presence of tactile paving from OSM data.
}
\details{
Note: the \code{osm_sf} must contain the tag: \code{c("tactile_paving")}.
}
\examples{
data = example_data
output = oi_im_tactile_paving(data)
plot(output$geometry)
}
