% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_inclusive_mobility.R
\docType{data}
\name{oi_inclusive_mobility}
\alias{oi_inclusive_mobility}
\title{Function to recategorize OSM data based on the recent Inclusive Mobility (IM) guide \href{https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/1044542/inclusive-mobility-a-guide-to-best-practice-on-access-to-pedestrian-and-transport-infrastructure.pdf}{UK}}
\format{
An object of class \code{function} of length 1.
}
\usage{
oi_inclusive_mobility(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap infrastructure data, obtained from the \href{https://github.com/ropensci/osmextract}{\code{osmextract}} function.
}}
}
\value{
The \code{osm_sf} data frame is returned with additional inclusive mobility columns returned.
}
\description{
Function to recategorize OSM data based on the recent Inclusive Mobility (IM) guide \href{https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/1044542/inclusive-mobility-a-guide-to-best-practice-on-access-to-pedestrian-and-transport-infrastructure.pdf}{UK}
}
\examples{
library(sf)
internal_data = extended_data
dim(internal_data)
output = oi_inclusive_mobility(internal_data)
}
\keyword{datasets}
