% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_is_lit.R
\docType{data}
\name{oi_is_lit}
\alias{oi_is_lit}
\title{Function to recategorise OSM infrastrcutre based on lighting presence.}
\format{
An object of class \code{function} of length 1.
}
\usage{
oi_is_lit(osm_sf, remove=FALSE)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data.
}}

\item{remove}{\itemize{
\item Boolean - If TRUE, features that do not have lighting
are removed from the \code{osm_sf} before being returned.
}}
}
\value{
an sf object with oi_is_lit column added, indicating if the way is
lit or not.
}
\description{
Determins whether or not a piece of OSM infrastructure is lit or not.
}
\details{
This function adds a new column, \code{oi_is_lit} containing values "yes"
, "no", whether or not the feature is lit. The value "maybe" is assigned when
there is a lack of data to distinguish - not implying the presence or lack of
lighting, but a lack of data to identify.

Note: the \code{osm_sf} must contain the following tags: \code{c("lit", "lit_by_led")}
}
\examples{
data = example_data
example_output = oi_is_lit(data)

# Quick Plot
plot(example_output["oi_is_lit"], key.pos = 1)

# Advanced plot with tmap - un-comment following four lines to run! 
#tmap::tmap_mode("view")
# tmap::tm_shape(example_output |> dplyr::select(oi_is_lit)) +
# tmap::tm_lines(col = "oi_is_lit", title.col = "Lighting presence") +
# tmap::tm_layout(legend.bg.color = "white")

}
\keyword{datasets}
