% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_road_names.R
\name{oi_road_names}
\alias{oi_road_names}
\title{Function to recategorise OSM data, combining both the \code{name} & \code{ref} field
for an OSM feature, or whichever is available.}
\usage{
oi_road_names(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A Simple Features \code{sf} and \code{data.frame} object containing
OpenStreetMap infrastructure data.
}}

\item{remove}{\itemize{
\item If TRUE, will remove all features that still have a NA for
openinfra_road_name column (removes nameless features).
}}
}
\value{
an sf object with openinfra_road_name column added, indicating the
name and ref fields of the feature, if they are included.
}
\description{
Function to recategorise OSM data, combining both the \code{name} & \code{ref} field
for an OSM feature, or whichever is available.
}
\details{
This function analyses OSM features, specifically the \code{name} and
\code{ref} fields that contain a road name (i.e. Otley Road) and the road
reference field (i.e A62). If both the \code{name} and \code{ref} appear, then
\code{openinfra_road_name} will be \code{name | ref}, otherwise \code{openinfra_road_name}
will be whichever field appears within the OSM data.

Note: the \code{osm_sf} must contain the following tags: \code{c("name", "ref")}
}
\examples{
data = example_data
data = data \%>\% dplyr::mutate(ref = "ref_field")
example_output = oi_road_names(data)
}
