% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_road_class.R
\name{recode_road_class}
\alias{recode_road_class}
\title{Re-classifies OSM Roads}
\usage{
recode_road_class(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap infrastructure data, obtained from the \href{https://github.com/ropensci/osmextract}{\code{osmextract}} function.
}}
}
\value{
The \code{osm_sf} simple features data frame is returned with additional columns road_class and road_desc based on Chan and Cooper's road classifications.
}
\description{
Determines road classification defined by \href{https://www.nature.com/articles/s41598-019-55669-8}{Chan and Cooper's} work.
Specifically re-classifies roads as one of 8 road classes (0-7) specified in the following \href{https://www.nature.com/articles/s41598-019-55669-8/tables/6}{table},
depending on each features \verb{highway=} key values.
}
\details{
Re-classifies data based on the following highway key values:\tabular{lll}{
   Road Class \tab Description \tab Selection from OSM \cr
   7 \tab Motorways \tab highway = motorway OR highway = motorway_link \cr
   6 \tab Non-residential Dual Carriageways \tab highway =  trunk OR highway = trunk_link*manual classification needed \cr
   5 \tab Residential Dual Carriageways \tab highway =  trunk OR highway =  trunk_link*manual classification needed \cr
   4 \tab Primary Roads \tab highway =  primary OR highway =  primary_link OR (highway =  trunk AND oneway = F) \cr
   3 \tab Secondary Roads \tab highway =  secondary OR highway =  secondary_link \cr
   2 \tab Tertiary Roads \tab highway = tertiary OR highway = tertiary_link \cr
   1 \tab Local Roads \tab highway = living_street OR highway = residential OR highway = unclassified \cr
   0 \tab Traffic-free Paths \tab highway = cycleway \cr
}
}
\examples{
library(sf)
u_data_large = paste0("https://github.com/udsleeds/openinfra/releases",
                      "/download/v0.2/bbbike_leeds_27_6_22.geojson")
u_data_small = paste0("https://github.com/udsleeds/openinfra/releases",
                      "/download/v0.2/30_06_22_bbbike_LCC_func_example_5_75km.geojson")
internal_data = extended_data
structure(internal_data)
output = recode_road_class(internal_data)
# Quick plot:
 plot(output["road_desc"], key.pos = 1)

# Advanced plot with tmap - un-comment following four lines to run! 
# tmap_mode("view")
# tmap::tm_shape(output |> dplyr::select(road_desc)) +
#  tmap::tm_lines(col = "road_desc", title.col = "Road class") +
#  tmap::tm_layout(legend.bg.color = "white")
}
