% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sotm_data}
\alias{sotm_data}
\title{OSM infrastructure data used in creation of transport infrastructure data packs.}
\format{
A data frame with 11,741 features described by 34 variables.
}
\source{
{\href{https://github.com/udsleeds/openinfra/blob/main/test-code/generate_example_data.R}{generate_example_data.R}}
}
\usage{
sotm_data
}
\description{
OSM infrastructure data used in creation of transport infrastructure data packs.
}
\examples{
library(sf)
sotm_data   # Lazy Loading. Data becomes visible as soon as it is requested. 
}
\keyword{dataset}
