% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{OSM infrastructure data to showcase the recode_road_class openinfra function.}
\format{
A data frame of infrastructure features for Leeds, with 36,065 features and 16 variables

total_place = osmextract::oe_get(\cr
place = place_name,\cr
provider = provider,\cr
layer = "lines",\cr
never_skip_vectortranslate = TRUE,\cr
force_download = TRUE,\cr
quiet = FALSE,\cr
extra_tags = c(all_extra_tags, "oneway", "maxspeed") # Add in "oneway" for recode_road_class)
}
\source{
{Created using the \href{https://github.com/ropensci/osmextract}{osmextract} package for R using oe_get() - params shown above.}
}
\usage{
openinfra::example_data
}
\description{
OSM infrastructure data to showcase the recode_road_class openinfra function.
}
\examples{
example_data   # Lazy Loading. Data becomes visible as soon as it is requested. 
}
\keyword{dataset}
