% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{OSM infrastructure data to showcase the recode_road_class openinfra function.}
\format{
A data frame of infrastructure features for Leeds, with 36,065 features and 16 variables

total_place = osmextract::oe_get(
place = place_name,
provider = provider,
layer = "lines",
never_skip_vectortranslate = TRUE,
force_download = TRUE,
quiet = FALSE,
extra_tags = c(all_extra_tags, "oneway", "maxspeed") # Add in "oneway" for recode_road_class)
}
\source{
{Created using the \href{https://github.com/ropensci/osmextract}{osmextract} package for R using oe_get() }
}
\usage{
data(example_data)
}
\description{
OSM infrastructure data to showcase the recode_road_class openinfra function.
}
\examples{
data(example_data)    # Lazy Loading. Data becomes visible as soon as it is requested. 
}
\keyword{datasets}
