% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{OSM ways data used in the creation of transport infrastructure data packs.}
\format{
A data frame of OSM ways.
}
\source{
{\href{https://github.com/udsleeds/openinfra/blob/main/test-code/generate_example_data.R}{generate_example_data.R}}
}
\usage{
example_data
}
\description{
OSM ways data used in the creation of transport infrastructure data packs.
}
\examples{
library(sf)
example_data   # Lazy Loading. Data becomes visible as soon as it is requested. 
}
\keyword{dataset}
