% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_active_cycle.R
\name{oi_active_cycle}
\alias{oi_active_cycle}
\title{Function to recategorise OSM infrastructure based on cycleability}
\usage{
oi_active_cycle(osm_sf, remove=FALSE)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} function.
}}

\item{remove}{\itemize{
\item Boolean - If TRUE, features that do not form part of a
cyclable network are removed from the \code{osm_sf} before being returned
}}
}
\value{
an sf object with oi_active_cycle column added.
}
\description{
Determines whether or not a piece of OSM infrastructure is cyclable or not
depending on whether the feature is included in the default cycling network
defined by \href{https://github.com/ropensci/osmextract}{\code{osmextract}}
}
\details{
This function mimics the cycling filtering steps applied by the
\href{https://rdrr.io/cran/osmextract/man/oe_get_network.html}{\code{oe_get_network}}
function, part of the
\href{https://rdrr.io/cran/osmextract/}{\code{osmextract}} package.

Note: the \code{osm_sf} must contain the following tags: \code{c("highway", "bicycle", "access", "service")}
}
\examples{
data = example_data
example_output = oi_active_cycle(data)
#plot(output["openinfra_cycle"], key.pos = 1)

}
