% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_clean_maxspeed_uk.R
\docType{data}
\name{oi_clean_maxspeed_uk}
\alias{oi_clean_maxspeed_uk}
\title{Re-classifies the maxspeed column of an OSM data frame to be compliant with current \href{https://www.gov.uk/speed-limits}{UK speed limits}. The clean re-coded speeds are stoed in \code{oi_maxspeed}.}
\format{
An object of class \code{function} of length 1.
}
\usage{
oi_clean_maxspeed_uk(osm_sf, no_NA = FALSE, del = FALSE)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap infrastructure data, obtained from the \href{https://github.com/ropensci/osmextract}{\code{osmextract}} function.
}}

\item{no_NA}{\itemize{
\item Boolean, \code{FALSE} by default. If \code{TRUE} then any oi_maxspeed == NA are removed. (i.e. maxspeed value is NOT compliant with \href{https://www.gov.uk/speed-limits}{UK speed limits}) are removed from \code{osm_sf}.
}}

\item{del}{\itemize{
\item Boolean, \code{FALSE} by default. If \code{TRUE} then the original \code{maxspeed} column is deleted and only \code{oi_maxspeed} is returned.
}}
}
\value{
The \code{osm_sf} simple features data frame is returned with the maxspeed column values cleaned based on \code{allowed_speeds}.
}
\description{
Re-classifies the maxspeed column of an OSM data frame to be compliant with current \href{https://www.gov.uk/speed-limits}{UK speed limits}. The clean re-coded speeds are stoed in \code{oi_maxspeed}.
}
\details{
Note: the \code{osm_sf} param must contain the following tags: \code{c("maxspeed", "highway")}
}
\examples{
library(sf)
internal_data = example_data
output = oi_clean_maxspeed_uk(internal_data, no_NA = TRUE)
plot(output["oi_maxspeed"])
 
#' # Advanced plot with tmap - un-comment following four lines to run! 
#' tmap_mode("view")
#' tmap::tm_shape(output |> dplyr::select(oi_maxspeed)) +
#' tmap::tm_lines(col = "oi_maxspeed", title.col = "Cleaned maxspeed") +
#' tmap::tm_layout(legend.bg.color = "white")
}
\keyword{datasets}
