% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_cycle_separation.R
\name{oi_cycle_separation}
\alias{oi_cycle_separation}
\title{Function to recategorise OSM data based on the type of cycle infrastructure.}
\usage{
oi_cycle_separation(osm_sf, remove=FALSE)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} function.
}}

\item{remove}{\itemize{
\item Boolean, FALSE by default, if TRUE, removes ways that are not
part of any dedicated cycling infrastructure.
}}
}
\value{
The provided \code{osm_sf} is returned with an additional column,
\code{openinfra_cycle_infa}, which indicates the type of cycling infrastructure
mapped as being one of the following values:
\verb{c("[cycle crossing](https://tinyurl.com/cycle-crossing)", "[cycle lane](https://tinyurl.com/cycle-lane)", "Mixed traffic - 20 mph" (no cycling infrastructure, carriageway maxspeed is 20 mph), "Mixed traffic - NA maxspeed" (no cycling infrastructure,  carriageway maxspeed value is NA), "Segregated - cycleway/track" ( infrastructure is segregated from carriageway by definition and is either a  [cycleway](https://tinyurl.com/higway-cycleway) or a  [cycle track](https://tinyurl.com/cycle-track)),  "shared lane/busway" (there is a lane drawn for cyclists, but it is either shared with a [busway](https://tinyurl.com/sahre-bus) or a  [shared lane](https://tinyurl.com/share-lane) with carriageway users)}
}
\description{
Function to recategorise OSM data based on the type of cycle infrastructure.
}
\details{
This function recetegorises OSM data on cycling infrastructure to one of the
protection categories defined within the LTN1/20 guidance (Mixed Traffic |
Cycle Lanes | Protected Cycling Space).

\strong{Note}: The \code{osm_sf} must contain the following tags: c("cycleway",
"cycleway_left", "cycleway_right", "cycleway_both")
}
\examples{
library(sf)
data = example_data
example_output = oi_cycle_separation(data, remove=TRUE)
plot(example_output$geometry) 
}
