% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_im_flush_kerb.R
\name{oi_im_flush_kerb}
\alias{oi_im_flush_kerb}
\title{Function to assess whether a kerb (where tagged) is flush or not.}
\usage{
oi_im_flush_kerb(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} package.
}}
}
\value{
The \code{osm_sf} data frame is returned with additional
\code{openinfra_im_flush_kerb} column.
}
\description{
Function to assess whether a kerb (where tagged) is flush or not.
}
\details{
Note: the \code{osm_sf} must contain the following tags: \code{c("kerb")}
}
\examples{
data = example_data
output = oi_im_flush_kerb(data)
output = output \%>\% dplyr::filter(! is.na(openinfra_im_flush_kerb))
plot(output[,"openinfra_im_flush_kerb"])
}
