% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_im_flush_kerb.R
\name{oi_im_flush_kerb}
\alias{oi_im_flush_kerb}
\title{Function to assess whether a kerb (where tagged) is flush or not.}
\usage{
oi_im_flush_kerb(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} package.
}}
}
\value{
The \code{osm_sf} data frame is returned with additional
columns \code{openinfra_im_kerb}, assessing the presence and type of kerb with
values \code{c("lowered kerb", "flush kerb", "no kerb", "raised kerb", "rolled kerb", "normal kerb", "unknown", "missing data")} and
\code{openinfra_accessible_kerb}, assesses accessibility of kerb for
pedestrians with impaired mobility with values \code{c("wheelchair - yes",  "wheelchair - no", "lacking data")} and \code{openinfra_im_flush_kerb} which
assesses whether or not there is a flush kerb compliant with the latest
\href{https://tinyurl.com/IM-guide-link}{Inclusive Mobility} Guidance.
}
\description{
Function to assess whether a kerb (where tagged) is flush or not.
}
\details{
Note: the \code{osm_sf} must contain the following tags: \code{c("kerb")}
}
\examples{
data = example_data
output = oi_im_flush_kerb(data)
output = output \%>\% dplyr::filter(! is.na(openinfra_kerb))
plot(output[,"openinfra_kerb"])
}
