% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_im_pedestrian_infra.R
\name{oi_im_pedestrian_infra}
\alias{oi_im_pedestrian_infra}
\title{Function to assess pedestrian infrastructure data.}
\usage{
oi_im_pedestrian_infra(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap
infrastructure data, obtained from the
\href{https://github.com/ropensci/osmextract}{\code{osmextract}} package.
}}
}
\value{
The \code{osm_sf} data frame is returned with additional
\code{openinfra_im_pedestrian_infra} column, containing descriptive values on
pedestrian infrastructure.
}
\description{
Function to assess pedestrian infrastructure data.
}
\details{
Note: the \code{osm_sf} must contain the following tags: \code{c( "footway", "sidewalk", "cycleway", "foot", "segregated", "access",  "crossing", "footway")}
}
\examples{
data = example_data
output = oi_im_pedestrian_infra(data)
plot(output["openinfra_im_footway"])
# Uncomment below to plot other columns
#plot(output["openinfra_im_footpath"])
#plot(output["openinfra_im_crossing"])
#plot(output["openinfra_im_footway_imp"])
}
