% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_inclusive_mobility.R
\name{oi_inclusive_mobility}
\alias{oi_inclusive_mobility}
\title{Function to recategorise OSM data based on compliance with the Inclusive Mobility (IM)}
\usage{
oi_inclusive_mobility(osm_sf)
}
\arguments{
\item{osm_sf}{\itemize{
\item A \code{sf} and \code{data.frame} object containing OpenStreetMap infrastructure data, obtained from the \href{https://github.com/ropensci/osmextract}{\code{osmextract}} function.
}}
}
\value{
The \code{osm_sf} data frame is returned with additional inclusive mobility columns returned.
}
\description{
This function is based on the Inclusive Mobility (IM) guide
\href{https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/1044542/inclusive-mobility-a-guide-to-best-practice-on-access-to-pedestrian-and-transport-infrastructure.pdf}{UK}.
}
\details{
See details in the link above, the function's source code, and our
justification
\href{https://udsleeds.github.io/openinfra/articles/im_get.html}{vignette}.

Note: the \code{osm_sf} must contain the following tags: \code{c("cycleway", "bicycle", "wheelchair", "kerb", "disabled", "mobility_scooter", "foot", "handicap", "lit", "access", "sidewalk", "footway", "incline", "width", "smoothness", "est_width", "sidewalk_left", "ramp_wheelchair", "priority", "footway_left", "sidewalk_both_surface", "footway_surface", "pedestrian", "footway_right", "ramp", "sidewalk_right", "sidewalk_both_width", "path", "sidewalk_left_width", "sidewalk_right_width", "sidewalk_right_surface", "sidewalk_left_surface", "maxspeed", "segregated", "sloped_curb", "surface", "tactile_paving")}
}
\examples{
library(sf)
internal_data = example_data
dim(internal_data)
output = oi_inclusive_mobility(internal_data)
}
